requirements.txt
aiofiles==24.1.0
annotated-types==0.6.0
anyio==4.2.0
Authlib==1.3.0
bonsai==1.5.2
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==42.0.4
dnspython==2.6.1
email-validator==2.1.0.post1
fastapi==0.115.6
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.2
httptools==0.6.4
httpx==0.26.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.3.2
natsort==8.4.0
argon2-cffi==23.1.0
pycparser==2.21
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
Pygments==2.18.0
pymongo==4.6.3
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.0
starlette==0.41.3
texttable==1.7.0
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.27.1
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
pyhiera
pyahocorasick==2.1.0